Adapt Keras' auto-config mechanism for inference/summary networks
This change makes it more capable for our purposes by allowing any
serializable value, not only the base types in the auto-config.
We have to check if this brings any footguns/downsides, or whether
this is fine for our setting.

It also replaces Keras' functions with our custom serialization
functions.