handle nesting: ConvertDType, ToArray, relax Concatenate
Concatenate can be equal to rename if only one key is supplied. By not
calling concatenate in that case, we can accept arbitrary inputs in the
transform, as long as only one is supplied. This simplifies things e.g.
in the `BasicWorkflow`, where the user passes the `summary_variables` to
concatenate, which may be a single dict, which does not need to be
concatenated.