minor adaptations to moving average (update time, init)
We should put the update before the standardization, to use the maximum
amount of information available. We can then also initialize the moving
M^2 with zero, as it will be filled immediately.

The special case of M^2 = 0 is not problematic, as no variance
automatically indicates that all entries are equal, and we can set
them to zero  (see my comment).

I added another test case to cover that case, and added a test for the
standard deviation to the existing test.