breaking: parameterize MVNormalScore by inverse cholesky factor
The log_prob can be completely calculated using the inverse cholesky
factor L^{-1}. Using this also stabilizes the initial loss, and speeds
up computation.

This commit also contains two optimizations.
Moving the computation of the precision matrix into the einsum, and
using the sum of the logs instead of the log of a product.

Open question:
Is the transformation behavior "left_side_scale" still correct for the
inverse matrix?

As the parameterization changes, this is a breaking change.
As it resolves major stability problems for higher-dimensional problems,
I thing it is worth including them anyway.