Support non-array data in test_quantity calibration ecdf [no ci]
Simulator outputs are allowed to be of type int or float, and
consequently have no batch dimension. This needs to be considered
in the broadcasting of inference_conditions for data based SBC
test quantities.

"examples/Linear_Regression_Starter.ipynb" contains an example where this is
necessary, where N is a non-batchable integer.