fix: layers were not deserialized for Sequential and Residual
As layers were passed with the `*layers` syntax, they could not be
passed as keyword arguments. In `from_config`, however, this was
attempted, leading to the layers to be ignored during reserialization.
This commit fixes this by taking the layers from `kwargs` if they are
passed as a keyword argument.