FusionNetwork: handle multiple networks with identical inputs
Up to now, the network strictly required passing the input to each
summary network, which required duplication of the data somewhere
upstream if the same data should be used for all backbones, which might
become a more common use case.

For this reason, I extended the fusion network with a second mode where
all backbones receive the same input. This is one possible
implementation, but we might also outsource this functionality into a
separate class.